Make publish resilient to pre-existing OAS's published to portal

+++ currency-v1/portal/publish.sh
@@ -51,12 +51,34 @@ export BAGGAGE_SPEC_ID=$(curl -X GET "http://kevinford-eval-test.e2e.apigee.net/
    -H 'X-Requested-With: XMLHttpRequest')
 echo "BAGGAGE SPEC ID IS:::::::::::::::::: $BAGGAGE_SPEC_ID"
 
-# DELETE EXISTING PUBLISHED BAGGAGE SPEC
-curl -X DELETE "https://e2e.apigee.net/portals/api/sites/kevinford-eval-boeing/apidocs/$BAGGAGE_SPEC_ID" \
+# GET EXISTING PUBLISHED FLIGHTS SPEC
+export FLIGHTS_SPEC_ID=$(curl -X GET "http://kevinford-eval-test.e2e.apigee.net/get-flights-spec" \
    -H "Authorization: Bearer $ACCESS_TOKEN" \
    -H "X-Org-Name: kevinford-eval" \
    -H "Accept: application/json, text/plain, */*" \
-   -H 'X-Requested-With: XMLHttpRequest'
+   -H 'X-Requested-With: XMLHttpRequest')
+echo "FLIGHTS SPEC ID IS:::::::::::::::::: $FLIGHTS_SPEC_ID"
+
+if [ "$BAGGAGE_SPEC_ID" != "0" ]; then
+    echo "FOUND BAGGAGE SPEC -- CLEANING";
+    # DELETE EXISTING PUBLISHED BAGGAGE SPEC
+    curl -X DELETE "https://e2e.apigee.net/portals/api/sites/kevinford-eval-boeing/apidocs/$BAGGAGE_SPEC_ID" \
+        -H "Authorization: Bearer $ACCESS_TOKEN" \
+        -H "X-Org-Name: kevinford-eval" \
+        -H "Accept: application/json, text/plain, */*" \
+        -H 'X-Requested-With: XMLHttpRequest'
+fi
+
+if [ "$FLIGHTS_SPEC_ID" != "0" ]; then
+    echo "FOUND FLIGHTS SPEC -- CLEANING";
+    # DELETE EXISTING PUBLISHED FLIGHTS SPEC
+    curl -X DELETE "https://e2e.apigee.net/portals/api/sites/kevinford-eval-boeing/apidocs/$FLIGHTS_SPEC_ID" \
+        -H "Authorization: Bearer $ACCESS_TOKEN" \
+        -H "X-Org-Name: kevinford-eval" \
+        -H "Accept: application/json, text/plain, */*" \
+        -H 'X-Requested-With: XMLHttpRequest'
+fi
+
 
 
 echo "OpenAPI Specification - BAGGAGE: Publishing Spec Apigee Edge Portal"
@@ -91,21 +113,6 @@ curl -i -X PUT "https://e2e.apigee.net/c3Rvc-ZG9j-1823/content" \
    -H "If-Match: $ETAG_FLIGHTS" \
    -H "Content-Type:application/x-yaml" \
    -d @currency-v1/portal/flights-spec.json
-
-# GET EXISTING PUBLISHED FLIGHTS SPEC
-export FLIGHTS_SPEC_ID=$(curl -X GET "http://kevinford-eval-test.e2e.apigee.net/get-flights-spec" \
-   -H "Authorization: Bearer $ACCESS_TOKEN" \
-   -H "X-Org-Name: kevinford-eval" \
-   -H "Accept: application/json, text/plain, */*" \
-   -H 'X-Requested-With: XMLHttpRequest')
-echo "FLIGHTS SPEC ID IS:::::::::::::::::: $FLIGHTS_SPEC_ID"
-
-# DELETE EXISTING PUBLISHED FLIGHTS SPEC
-curl -X DELETE "https://e2e.apigee.net/portals/api/sites/kevinford-eval-boeing/apidocs/$FLIGHTS_SPEC_ID" \
-   -H "Authorization: Bearer $ACCESS_TOKEN" \
-   -H "X-Org-Name: kevinford-eval" \
-   -H "Accept: application/json, text/plain, */*" \
-   -H 'X-Requested-With: XMLHttpRequest'
 
 echo "OpenAPI Specification - FLIGHTS: Publishing Spec Apigee Edge Portal"
 # PUBLISH THE BAGGAGE SPEC
